Make structure of packages more consistent with others
This was requested here:
<https://github.com/zigpy/zha-device-handlers/pull/2928#issuecomment-1913802591>
